[TensorExpr] Pull changes to core classes for representing expressions and statements from the side branch.
Our development has been happening on a side branch `pytorch_fusion` in
`bertmaher/pytorch` fork. This PR moves changes to the core classes
representing expressions and transformations on them.

At this moment, the tensor expressions are only used in tests.
Subsequent PRs add LLVM and CUDA codegen for tensor expressions and
implement fuser on top of these.

This PR is huge as it is a squashed version of changes in the side
branch. It is not practical to pull changes one by one from the branch,
so here is the squashed version. If you're interested in seeing the
history of changes, please refer to https://github.com/bertmaher/pytorch

ghstack-source-id: 1b0fdf2cc2499f62f54965ac1c68a080241a3322
Pull Request resolved: https://github.com/pytorch/pytorch/pull/34224